requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
greenlet==3.2.3
markdown-it-py==3.0.0
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
rich==14.0.0
sentry-sdk==2.33.2
shellingham==1.5.4
SQLAlchemy==2.0.41
typer==0.16.0
typing_extensions==4.14.1
urllib3==2.5.0
